+++ src/Latest Backend Project/TeamPhoenix.MusiCali/TeamPhoenix.MusiCali.Services/BingoBoard.cs
@@ -5,7 +5,7 @@ namespace TeamPhoenix.MusiCali.Services
 {
     public class BingoBoard
     {
-        public static List<GigSummary>? ViewMultGigSummary(ushort numberOfGigs, string username)
+        public static List<GigSummary>? ViewMultGigSummary(int numberOfGigs, string username)
         {
             try
             {
